Move fetch methods to fetch file, rename file
